Fix windows build error.
git-svn-id: svn://anonsvn.kde.org/home/kde/branches/work/kst/portto4/kst@963421 283d02a7-25f6-0310-bc7c-ecb5cbfe19da

+++ src/libkstapp/datawizard.cpp
@@ -903,8 +903,7 @@ void DataWizard::finished() {
 
     Q_ASSERT(plotItem);
 
-
-    double fontScale = ApplicationSettings::self()->referenceFontSize()/sqrt(plotList.count()) - ApplicationSettings::self()->referenceFontSize()+ApplicationSettings::self()->referenceFontSize()/3;
+    double fontScale = ApplicationSettings::self()->referenceFontSize()/sqrt((double) plotList.count()) - ApplicationSettings::self()->referenceFontSize()+ApplicationSettings::self()->referenceFontSize()/3;
     if (fontScale>0) fontScale = 0;
 
     plotItem->setGlobalFontScale(fontScale);
